276 / Use dropdowns for role management (#295)
Fixes #276 

 - Convert role toggles for permissions management to dropdowns
 - Rewrite role management handler to add more robust edit policies
 - Clean up `Toggle` component

For future reference:

The glaring issue with using a dropdown here is that roles are not
actually mutually exclusive entities in the authorization model. So,
it's possible that a user can be assigned to multiple roles, and now the
UI is more limited in expressing that circumstance.

While there may not be much of a reason to explicitly assign people to
multiple roles now (since roles are basically hierarchical), it will
still come up as a point of confusion due to the way permissions can be
inherited. For example, an admin from an institution inherits admin for
a class, which can neither be revoked in this UI nor even expressed very
clearly. (I added some limited way of handling this case, but since it's
generally not possible to distinguish inherited vs. assigned
permissions, there's only so much we can do.)

In the future we might also choose to add roles with disjoint
capabilities, meaning we'd need to bring back some sort of multiselect /
toggle for assigning roles.

All that said, for the time being, this change will make permissions
management more intuitive for everyone (except me).