Fix loading on main thread page (#298)
- Main thread page loading spinner was buried in the stacking order
 - Show loading spinner on navigation changes
 - Center text that's shown on various no data states

There's a small inconsistency now between the loading screen on the
thread page and on the landing page where on the former the chat input
is left uncovered by the modal backdrop, while on the latter the chat
input is covered. Not sure which is right, so we'll wait and see who
complains about the inconsistency to find out which it should be.