Fix db init script (#300)
Fix a bug @kylelarkin encountered with the `start-dev-docker.sh` script.

The `db init` command was aggressively stamping the database with the
latest DB migration version, even if it had not applied the latest
migrations. This would cause the `db migrate` command to do nothing when
trying to apply future migrations because it thought it was at the
latest version.

Fixes:
- Revise `db init` to do nothing at all if the DB already exists, unless
`clean` is explicitly requested (in which case we drop, recreate, and
stamp at head, which will be correct in that case)
- Run `db migrate` after `db init` in the `start-dev-docker.sh` script
in order to apply new migrations and get the DB into the latest revision
state the right way.