304 / Fix init script for clean installs (#313)
Fixes #304 

FYI @ekassos 

The new `blank_slate` check assumed (incorrectly) that if the database
schema existed there would be tables in it / a brand new database would
not have the database schema in it. But postgres actually does create
the right schema when initializing the container, without any tables in
it.

Change the check to 1) always ensure the schema exists / create it if
necessary, then 2) see if there are any tables in the schema and infer
it is a `blank_slate` only if there are no tables yet.

Now the script should initialize a brand new DB correctly and also
migrate existing DBs correctly with no data loss.