Migrate to Assistants API v2 (#303)
This pull request closes #222 by migrating to v2 of the OpenAI
Assistants API.

## Background: Main Changes to OpenAI Assistants API

- The `retrieval` tool has been renamed to the `file_search` tool
- Files are now associated with tools instead of Assistants and
Messages.
- Files are attached to Assistants and Threads using the new
`tool_resources` object.
- Threads only support the tools that the Assistant has been configured
with.
- The `tool_resources` for the `code_interpreter` tool are a list of
`file_ids`.
- The `tool_resources` for the `file_search` tool are now a pre-created
`vector_store`. Vector stores take a list of `file_ids`.
- **One `vector_store` per Assistant and/or Thread.** We can end up
with:
  - No `vector_store`.
- One `vector_store` attached to the Assistant and no `vector_store`
attached to the Thread.
- One `vector_store` attached to the Thread and no `vector_store`
attached to the Assistant.
- One `vector_store` attached to the Assistant and a separate
`vector_store` attached to the Thread.
- There are helper functions to create a `vector_store` when creating a
thread, or upload files to a thread's `vector_store` or Code Interpreter
`tool_resources`. We do not use any of those helper functions. See *A
note on the expiration of `vector_store`s*.

## Design/UI changes

- Assistants Edit/Create page:
- The Tools section now features two checkboxes for File Search and Code
Interpreter. The original toggle version was scrapped for a more
consistent look with the rest of the input fields on the form. Also note
that `Toggle.svelte` was removed in
870064a42f549d3ba924b505a1b6bffd565ce96c.
  - Separate File Search and Code Interpreter file multiselects.
- ChatInput:
- Two separate buttons for File Search and Code Interpreter file
uploads.
- Buttons are enabled only when the associated assistant has those
capabilities turned on. Note that tools can not be enabled on the Thread
side, and need to be managed on the Assistant side. This is an API
limitation; providing `tool_resources` without the tool enabled creates
inconsistent behavior. For example, the model might generate code using
Code Interpreter, but it cannot run the code.
- FileUpload:
- File uploads in the Thread page are added in two different piles.
Might be worth it to investigate a different approach for file uploads.
See *Future Directions* below.

## Database changes
- There is now a Vector Stores table. 
- Assistants now directly map to Code Interpreter files only. Assistants
map to one or no Vector Stores.
- Threads now directly map to Code Interpreter files. Threads map to one
or no Vector Stores. `tools_available` is used to keep track of `tools`
the Thread has access to. See *Future Directions* for more discussion.

## A note on the expiration of `vector_store`s
Vector stores created using thread helpers (like
[tool_resources.file_search.vector_stores](https://platform.openai.com/docs/api-reference/threads/createThread#threads-createthread-tool_resources)
in Threads or
[message.attachments](https://platform.openai.com/docs/api-reference/messages/createMessage#messages-createmessage-attachments)
in Messages) have a default expiration policy of 7 days after they were
last active (defined as the last time the vector store was part of a
run). When a vector store expires, runs on that thread will fail.

Because we do not use these functions and take on a manual
`vector_store` creation, our `vector_store`s never expire. See *Future
Directions* for more discussion.


## Breaking changes

Because of the DB changes, old Threads and Assistants will not work
without a migration script. However, this change should be quite simple
as the old Assistants are mapped to the new Assistant object format when
calling `GET` on `https://api.openai.com/v1/assistants`:

```json
{
    "id": "asst_kO8I7e5b1004CCrgYLPxsd4L",
    "object": "assistant",
    "created_at": 1717731682,
    "name": null,
    "description": null,
    "model": "gpt-4-1106-preview",
    "instructions": "Be very happy.\n---Other context---\nThe current date and time is 2024-06-07 03:41:22 (UTC).",
    "tools": [
        {
            "type": "file_search"
        },
        {
            "type": "code_interpreter"
        }
    ],
    "top_p": 1.0,
    "temperature": 1.0,
    "tool_resources": {
        "file_search": {
            "vector_store_ids": [
                "vs_2qVugxNW4Bla1RCQUOp1AKIv"
            ]
        },
        "code_interpreter": {
            "file_ids": [
                "file-NP9dpQvVu5n6BiMf0jbwQNpU",
                "file-HITTZTLvgtxdXuOHxYGfKWLJ"
            ]
        }
    },
    "metadata": {
        "class_id": "1",
        "creator_id": "1"
    },
    "response_format": "auto"
},
```
## Future directions

Some things that can/should be investigated/resolved as separate issues,
since they are out of scope of this issue:
- Experiment with versions of the file upload UI for Threads. Now that
we have access to all files relevant to the thread, we could offer an
overlay that allows adding/deleting files instead of the blocks we're
using now. Note that file counts can get pretty large in this new
version. Another potential exploration is different file icons for the
two types, although it could be confusing to users.
Also note that we have no visibility of uploaded thread files after
sending the respective message. This might lead to double-uploading or
user confusion. See #284
- Validate counts for Thread file uploads. Since we are currently only
validating this at the `CreateThread`/`NewThreadMessage` level, we
cannot have extra dynamic file count validation. See #301
- Synchronize Assistant and Thread available `tools`. Because Threads
currently decide on their available tools during creation, a change in
Assistant configuration might create a situation where a Thread calls on
an unavailable tool. See #302
- Add more control for Assistant: `top_p`, `temperature`. See #305
- Fix the Code Interpreter message persistence issue. From my testing,
it looks like the Code Interpreter output is appended to the previous
Assistant message, while the Assistant generates a new text message. See
#240
<img width="1694" alt="image"
src="https://github.com/stanford-policylab/pingpong/assets/55115777/d674455d-3782-4f52-ae0a-cf78b037f135">