306 / Migrate DB after migrating to v2 of Assistants API (#321)
Provides the script for the DB migration after applying the DB changes
in #303, and closes #306.

## Testing

### Assistants
Verified the correct migration of the following types of assistants by
creating them before implementing the database schema:

| `assistant_id` | File Search | Code Interpreter | Uploaded Documents*
|
| :---  | :---:  | :---:  | :---:  |
| `v1_file_only_no_uploads`  | ✓ | X | X
| `v1_file_code_no_uploads`  | ✓ | ✓ | X
| `v1_file_only_with_uploads`  | ✓ | X | ✓
| `v1_file_code_with_uploads`  | ✓ | ✓ | ✓

*\* on the Assistant edit page*

All assistants migrated successfully, and editing the assistants after
the migration was also successful.

### Threads

Verified the correct migration of the following types of threads by
creating them before implementing the database schema:

| `thread_id` | File Search^ | Code Interpreter^ | Uploaded Documents† |
Attachments‡ |
| :---   | :---:  | :---:  | :---:  | :---:  |
|`v1_file_only_no_uploads_no_attachments` | ✓ | X|X |X |
|`v1_file_only_no_uploads_with_attachments` |✓ |X | X| ✓|
|`v1_file_code_no_uploads_no_attachments` | ✓ | ✓|X | X|
|`v1_file_code_no_uploads_with_attachments` | ✓ |✓ | X| ✓|
|`v1_file_only_with_uploads_no_attachments` | ✓ |X | ✓|X |
|`v1_file_only_with_uploads_with_attachments` |✓ |X | ✓|✓ |
|`v1_file_code_with_uploads_no_attachments` | ✓ |✓ | ✓|X |
|`v1_file_code_with_uploads_with_attachments` | ✓ | ✓| ✓| ✓|

*^ determined by the Assistant-enabled capabilities*
*† on the Assistant edit page*
*‡ uploaded on the Thread*

All threads migrated successfully, along with any images and graphs
generated in each thread. Access to attachments was verified after
migration. The capabilities of each assistant (file search/code
interpreter) were correctly reflected in ChatInput. Note that the issue
in #240 is not resolved.

---------

Signed-off-by: Evangelos Kassos <github@evangeloskassos.com>