52/ Assistant Deletion (#395)
Closes #52 by introducing support for deleting assistants.

### Other changes & fixes:
- `canSubmit` is passed along to `ChatInput` to display custom message
when viewing another user's thread: "You can't reply in this thread"
- `loading` screen now properly overlays the Edit Assistants page before
navigating to the All Assistants view.
- In some edge cases where a file is deleted from an OAI VectorStore but
the DB operation is rolled back, a `NotFoundError` might be returned the
following time a deletion of the same file is attempted. This error is
resolved by ignoring `NotFoundError`s in VectorStore and File deletion
operations.