Support for GPT4o Mini & Model Selector Overhaul (#402)
Adds support for GPT-4o Mini and a new Model Selector design. There are
now "New", "Recommended" and "Vision capabilities" badges available for
each model, along with the existing model description. This new design
should help users not familiar with the models make an informed decision
on which one to use.

Note that the "Vision capabilities" badge will only appear when Vision
is enabled across the app.