Model Selector tweaks & Dropdown refactoring (#407)
This PR adds design updates to the Model Selector and refactors the
selector elements into standalone components so that they may be used in
other parts of the interface, especially for immediate use in #405.

### New components:
- **`DropdownContainer`**: Contains the Button and DropDown component
that gets triggered when the button is clicked.
- **`DropdownHeader`**: Colored sticky section headers for DropDown
component. `topHeader` introduces a spring effect and should be used for
the first header of the list.
- **`DropdownFooter`**: Always-showing footer at the bottom of the
DropDown list. Can be used with a link or other components.
- **`DropdownOption`**: Main Dropdown Option component, which can
include `DropdownBadge`s. Two options for title styling, with a subtitle
option.
- **`DropdownBadge`**: Badges to be included in the `DropdownOption`.
They are typically placed on the right end of the `DropdownOption`'s
title row.
- **`ModelDropdownOptions`**: Wrapper for generating `DropdownOption`s
based on `AssistantModelOptions`.

### Design tweaks:
- The dropdown headers (Latest Models, Pinned Models) have a colored
background to better stand out on the list.
- The dropdown headers stick at the top of the list while users scroll
through the applicable dropdown items.
- The top dropdown header has a spring-like effect that extends the
dropdown header background color as users scroll past the top of the
list for better responsiveness.
- The Model Selector has a footer directing users to OAI's model
descriptions.
- Badges for Dropdown Options have a gradient effect to better stand out
against the selected/hovered background of the Dropdown Options.
- The Vision/No Vision capabilities badge next to the Model Selector has
been updated with a background gradient for consistency.