New Edu Access, Privacy Policy pages, revamped About page (#416)
Closes #415 by introducing the Privacy Policy and Educational Access
pages, and revamping the About page. All three pages are now available
without login.

## Component changes:
- **NEW:** `PingPongDemoCarousel` includes four screenshots of the
PingPong interface and is used in the Educational Access and About
pages.
- **UPDATED**: `Sidebar` now displays a Login link in the slot of the
"Start new chat" button when the user is not authenticated. In the same
circumstances, the Recent threads and Profile `SidebarGroup`s are
hidden, and replaced with links to "Home" (About), "Educational Access,"
and "Privacy Policy."

## Other changes and fixes:
- `pingpong.hks.harvard.edu` now redirects to
`pingpong.hks.harvard.edu/about` instead of
`pingpong.hks.harvard.edu/login`. This change should give non-users a
better overview of what PingPong is compared to the login page.
- Fixes some small typos in non-related files.