Canvas Sync: Option for other users to disconnect an account (#473)
Allows users whose Canvas account is not currently connected to the
PingPong group to remove said connection and add their own. Presents
option to keep or remove users that have been imported through Canvas.

Also fixes an error introduced in #463 where calling `DELETE` on the
Canvas OAuth2 endpoint on disconnect would result in a `401
Unauthorized` error if the token was already expired.