Canvas Sync: Merge duplicate accounts based on SSO (#474)
Merges accounts based on SSO identifiers already known for each user.
Few changes to make that happen:
- There's a new `external_logins` table that stores known SSO
identifiers for each user.
- When the user logs in through SSO, we try to find their email or SSO
identifier in an account to avoid creating a duplicate. This is a
two-step check as we want to make sure we don't create orphaned
email-matching accounts by logging into the SSO-matched account. We also
update the email address to match the SSO email address.
- When a user is synced through Canvas we perform the same checks.
- There's a new `MergeUsers` class handling the merge.
- There's a new `users_merged_users` table used to identify users who
have been merged, as we do not update message metadata on OAI's end.
- There's a new `users_merged_users` table used to identify users who 
have been merged, as we do not update message metadata on OAI's end.