Remove more class admin things (#493)
Fixes #488 

In this PR:
- Removes "Administrator" from dropdown in the UI unless the person
still has this role
- Works towards #488 by updating permissions model to move the
`can_manage_assistants` permission over to `supervisor` instead of
`admin`.
- Allow testing on functions that write / delete permissions with the
authz mock server
- Allow inspection of permissions operations that occurred during the
test on the authz mock server
- Add new tests for the group creation endpoint to verify correct
permissions setting on public vs. private groups.
- We will no longer grant `admin` automatically to the creator of a
group.
- Note that `admin` role is required to set the API key. Since the
creator of the class had to have had `admin` anyway this shouldn't be an
issue.
 - Script for migrating permissions.


Migration steps:
- [x] Upgrade OpenFGA to 1.6.0 (this is actually optional but a good
idea)
- [ ] Run `python -m pingpong auth update_model` to update the latest
permissions model
- [ ] Run script to clean up permissions: `python -m pingpong auth
update_group_admin_perms`
- [ ] Finally, the permissions model can be updated again after the
migration is complete to remove the `class#admin` userset on
`can_manage_assistants`. This is needed during the migration in order to
check this permission and convert it to supervisor.
 - [ ] Restart pingpong srv to pick up newest model