Fix: Users might not be able to disconnect Canvas accounts from other users (#492)
Fixes an issue where users might receive error "You're not the
authorized Canvas user for this class" when trying to disconnect the
Canvas accounts of other user linked with a group. Specifically, a check
of whether the user making the request is the user whose Canvas account
is connected for the class would fail. This check is no longer enforced
when we are making a `log_out()` call.