New workflow to push images on staging from PRs (#518)
This new workflow allows for testing on PR images without the need to
push to `main`. This workflow pushes a `stage` and `pr_###` on the
new-built image. When we push a different image with the same `stage`
and `pr_###`, the old image remains untagged. A new lifecycle policy in
the ECR repository collects and deletes those untagged images.

To trigger the workflow, type `/test-on-staging` in a PR comment.