Regenerate title for chat thread when we have more information (#553)
Closes #438 by setting the thread name when we have more information. In
particular, the thread name will be updated right before we add the
third user message to the thread. At that time, we should expect to have
at least 3 user and 2 assistant messages, which should provide enough
context for a relevant conversation title.

### Notes
- **Deployment:** DB schema migration needed.
- Note that since we do not update `Sidebar` when we send a message, the
new thread title will only appear when the user refreshes the page. The
same applies for the timestamp and list ordering. See #552.
- The new default thread title before 3 user messages are sent is "New
Conversation."