Add New Users workflow revamp (#554)
Closes #287, #289, #491 by revamping the Add New Users workflow. Users
are now able to use any of the accepted RFC5322 email formats when
adding users to PingPong. We now also allow users to "cure" any invalid
email addresses on the frontend before submitting a final request.
Finally, we do not invalidate the whole request when an error occurs but
communicate the error at a final summary page.