Fix: Users should not be able to interact with assistants they no longer have permissions to view  (#561)
Reinforces requirement both server side and client side that users
should have the `can_view` permission for the assistant associated with
the thread to reply to a thread. Otherwise, threads become read-only.

This PR also resolves an issue where a non 200 response would be
displayed as an `[Object]` error. The error message now displays the
actual error `detail`.