Fix: Thread-related Vector Stores are not attached to thread (#573)
Fixes an issue where empty thread vector stores we created were not
attached as tool resources to the relevant thread. This resulted in OAI
creating a second vector store with a set (7 day) expiration date. A
side effect of this issue is that we were keeping track of the wrong
vector store, which might result in incomplete thread deletions.