Canvas Sync: Add SIS viewing permissions check (#610)
Closes #605 by introducing a permissions check right before users select
a class to connect with Canvas. This check is introduced in
`verify_access` under the `canvas` module, and is run twice: once when
the user selects the class, and once when they save the class.

If `missing_sso_ids` is true when syncing, the user will need to
reconnect their account.