Azure API Deployment-related Bug Fixes and Improvements (#633)
Fixes the following issues:
- Creating an Azure OpenAI client fails because of a misnamed parameter
being passed in the `openai.AsyncAzureOpenAI` call.
- When trying to set an API key, the provider is always set as `openai`,
even when the user selected Azure as the AI Provider.
- The Azure OAI Client model objects have a `created_at` instead of a
`created` field, and no `owned_by` field.
- The Azure client /list endpoint returns different available model
names than the OAI client.

Also made the following improvements:
- Classes classes included the API key parameter in every call. Removed,
and introduced an `api_key/check` endpoint to further limit transfer of
API key data between the server and the client.
- Saving an assistant with an unavailable deployment name shows an
unhelpful generic error. Now, users are shown an error with the required
deployment name.

## Known Issues
- PingPong cannot verify the capabilities of an Azure model deployment
using the Azure client. Therefore, we rely on the user to create
deployments that correspond to the model name. For example, PingPong
will assume that deployment "gpt-4o-mini" corresponds to model
"gpt-4o-mini" with Vision capabilities, but it can be the deployment of
any OAI model.
**No workaround known.**
- Azure model deployment names need to *exactly* match the model name
provided by the /list endpoint. PingPong will generate an error if the
deployment name cannot be found in the Azure endpoint provided.
**No workaround known.**