Prevent Edits on Locked Assistants (#654)
Adds the ability to prevent edits on locked assistants. The Manage
Assistant page is renamed to View Assistant. There is a new
`/class/{class_id}/assistant/{assistant_id}/prevent_edits` endpoint for
toggling the locked status of the assistant. When an assistant is locked
to prevent edits, it cannot be deleted, and all fields besides the
Publish button are disabled. Only Group (Institutional or Root) Admins
can change the locked status of an Assistant.