Support multiple email addresses under the same user account (#667)
This PR request introduces support for emails as an additional
ExternalLogins attribute in a User object. Additionally, it modifies the
`/login/magic` endpoint to prioritize checking the request’s email in
ExternalLogins, in addition to the `User.email` field.

## Motivation
There may be situations where we might need to identify a user under
multiple emails. For example, users might have trouble receiving emails
from PP's email account address in their institutional addresses.

## Implementation
This feature is implemented as an endpoint accessible to administrators
and is not present in the UI. The `/user/add_email` endpoint accepts an
`AddEmailToUserRequest` with `current_email` and `new_email` fields. If
the `new_email` is already used in another account as a user `email` or
`ExternalLogin`, the endpoint raises an error.

## Next Steps
- Add the option to add additional emails for a specific user in the UI.
This step will require ensuring successful merging of accounts.
- Migrate all current user emails as new ExternalLogins.
- Display all ExternalLogins in the User List under Manage Group and the
User's profile. Otherwise, determine which email to display in the User
List under Manage Group.