Fix: APIKey Radio Selector causes Viewport jump (#690)
Resolves #688 in Chromium browsers. This is expected Chromium behavior:
https://issues.chromium.org/issues/40782006. Webkit does not behave in
this way, so the issue doesn't affect Safari users.

Adapted from
https://github.com/twbs/bootstrap/issues/39097#issuecomment-1867954762,
but note that we don't have a prop to add this change to, so we have to
use custom styles.

> [Radio
styling](https://flowbite-svelte.com/docs/forms/radio#Radio_styling)
> Use the class prop to overwrite the label and input tag class.

Adding `hidden` to the `class` prop changes both the label and input tag
classes, which hides the actual label.