Fix: Properly display callback errors (#703)
Fixes an error where an error message would not display above
`ChatInput`. This was caused by `form.callback` accepting an
`CallbackParams` object that was not properly unpacked.

### Other fixes:
- Removes an old `rate_limit_error` type that the current thread manager
does not know how to handle and replaces it with the correct
`presend_error` type that returns the typed message and input to
ChatInput.
- Moves the catcher for `openai.APIError` to the right place to catch
errors when creating a thread.
- Adds a second `openai.InternalServerError` with more information for
the user.