New merge endpoint and `merge_external_logins` logic fix (#707)
Adds a new `/user/merge` endpoint so that we don't have to use a CLI
command every time we need to merge two users, which is increasingly
common in the RCT.

Also fixes a logical error in the `merge_external_logins` function.
Previously, ExternalLogins with uniqueness constraints across all users
(for instance, an email can only be used as an ExternalLogin for a
single user) would raise an error. This was because the old function
first created the new ExternalLogin record and then deleted the old one.