[Canvas Sync] Handle Missing Emails or Invalid Tokens (#722)
- Adds new `CanvasInvalidTokenException`, which is raised when we cannot
receive a refreshed token through `/login/oauth2/token`. When there is a
`ClientResponseError`, `with_retry` will force refresh the
authentication token.
- Adds support for raising custom errors for specific
`ClientResponseError`s in `with_retry`. We do not want to specify this
within `with_retry` to keep it generic enough.
- Canvas Sync will now check that we receive emails and enrollment
information for users when verifying access to course rosters and before
attempting to create `UserClassRole`s when syncing. Classes will be
marked as having a sync error in both cases.
- Handle log outs more gracefully: If we don't find the token
(`CanvasInvalidTokenException`) we can ignore the error.
- New `Spinner`s in the UI to account for retries when logging out from
Canvas.