Support for Reasoning Models & Better Assistant View File Management (#726)
Adds support for Reasoning models, including `o1` and `o3-mini`. Users
can specify a `reasoning_effort` to guide the model. We are unable to
display reasoning output tokens, as they are not returned by the API.

`o1` and `o3-mini` are currently only generally available in OAI
deployments, so we are not displaying them in the model selector for
Groups using Azure as their AI Provider.

This PR also includes enhancements in the management of files in the
Assistant view. Files users upload will now be deleted if the users
navigate away from the page before saving.