Support for User Agreements (#753)
Adds support for Agreements, Agreement Policies, and storing timestamps
for when users agree to the rules. Agreement Policies can be "assigned"
to the entire user base, or to a subset of users who have at least one
ExternalLogin with a specific provider.

Resolves an issue in `parse_session_token`, where after changing the
verification status of a user, the user object was not refreshed. Also
resolves an issue in `test_server.py/test_me_with_valid_user` where we
expected an `updated: None` property in the User object after changing
the verification status of the user.