Do not allow users to select admin-only or hidden models (#783)
Builds on #782 and #781 by adding server-side checks that a model being
used in an assistant is not hidden, or that the user is an admin when
trying to select an Admin-only model. This change does not affect
requests which do not attempt to change the assistant model. For
example, when a non-Admin user tries to change any other properties of
an assistant with an Admin-only model set, the changes will be saved
successfully.