Log messages that trigger Azure content filtering (#792)
We are seeing `BadRequestError`s being raised when Azure content filters
do not allow the model to return a response. Logs the messages
transcript for us to understand what messages are triggering the filters
(falls under the following clause in the Privacy Policy):

> The Harvard technical team developing and managing PingPong does have
access to identifiable information for purposes of system
administration, including system protection, maintenance, improvement,
and management as well as **abuse prevention**, to track user trends to
improve the quality of PingPong and the learning experience, and for the
other purposes described in this notice and other relevant policies.