[CanvasSync] Keep the latest active profile when there are duplicates (#794)
Adds a new `last_active` parameter in `CreateUserClassRole` to track
when each Canvas profile was last active. This is an improvement over
#791, where the student profile that appeared latest in the list would
be the one to persist the SSO information.

As a reminder, duplicate profiles are defined as ones that use the same
email address in Canvas. In this approach, we check which of the
duplicates was used most recently and save the SSO identifier of that
profile.